Modified version of sudoku1.py
Modified version of sudoku1.py showing how the same algorithm with a trivial
optimization gives a huge performance increase.

A more comprehensive algorithm will definitely be faster and better at solving
the harder problems, I just think it is useful to show how much faster the
baseline solver can be :-)

The main reason this is so much faster is that sudoku1 calls the same_ functions
millions of times to solve a puzzle:

    4030560 same_row
    3582720 same_col
    3184640 same_block

There are only 6561 different combinations that need to be checked, and
pre-computing the results removes the need for 10,000,000 function calls.